+++ include/ui/ui.h
@@ -21,13 +21,19 @@ struct rui_draw_data_t {
   size_t vertexCount;
 };
 
+extern struct rui_context_t uiCtx;
+
+/**
+  @brief Clears uiCtx
+*/
 enum reh_error_code_e rui_Begin(struct rui_context_t *uiCtx);
 enum reh_error_code_e rui_SetupRenderData(GLuint *program, GLuint *VAO, GLuint *VBO, GLuint *EBO);
-enum reh_error_code_e rui_RenderRect(struct rui_context_t *uiCtx, float x, float y, const char* label, float w, float h, struct rm_vec3_t color);
+enum reh_error_code_e rui_RenderRect(struct rui_context_t *uiCtx, float x, float y, const char* label, float w, float h, struct rm_vec3_t color, void (*onHover)(struct rui_command_rect_t *), void (*onClick)(struct rui_command_rect_t *));
 enum reh_error_code_e rui_End(struct rui_context_t *uiCtx, GLuint *program, GLuint *VAO, GLuint *VBO, GLuint *EBO, float **projectionMatrixPtr);
 
 enum reh_error_code_e rui_DrawRect(float x, float y, float w, float h, struct rm_vec3_t color, struct rui_draw_data_t *drawData);
 
 enum reh_error_code_e rui_AABBCollisionCheck(struct rui_command_t *A, struct rui_command_t *B, bool *collision);
+enum reh_error_code_e rui_AACursorCheck(int *id);
 
 #endif // RUI_H